Update pollen service url
Pollen service was updated.  Took this as an opportunity to adjust how we are proxying this information.  We now proxy using an AWS Lambda python call via API gateway -- before we were just trying to proxy with API gateway.  This makes future maintenance a bit more straightforward and allows for easier inclusion of failover urls.